requirements.txt
fastapi
uvicorn
pandas
python-multipart
python-dotenv
requests
numpy
sentence-transformers
pydantic
